refactor : Service Interface 제거
Project/Epic/Task Interface 제거
