+++ src/main/kotlin/com/dodo/dodolistserver/epic/EpicController.kt
@@ -3,10 +3,12 @@ package com.dodo.dodolistserver.epic
 import com.dodo.dodolistserver.common.dto.ResponseDto
 import com.dodo.dodolistserver.common.message.ResponseMessage
 import lombok.RequiredArgsConstructor
+import lombok.extern.slf4j.Slf4j
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
+@Slf4j
 @RestController
 @RequestMapping("/api/v1/epics")
 class EpicController (private val epicService: EpicService) {
@@ -21,7 +23,6 @@ class EpicController (private val epicService: EpicService) {
     @GetMapping(path = ["/{epicId}"])
     fun getEpicByEpicId(@PathVariable("epicId") epicId: Long): ResponseEntity<Any> {
         val selectedEpic = epicService.getEpicByEpicId(epicId)
-
         return ResponseEntity.ok()
             .body(ResponseDto.of(HttpStatus.OK, ResponseMessage.SUCCESS, selectedEpic))
     }
